Simplify move generator classes
