+++ test/movegenerator_test.cpp
@@ -18,11 +18,11 @@ TEST(MoveGeneratorTest, CandidateMoves1)
 {
    Board b;
    State s;
-   MoveGenerator gen(b,s);
+   std::vector<Move> moves;
+   MoveGenerator gen(b,s, std::back_inserter(moves));
    MoveExecutor exec(b, s);
 
-   std::vector<Move> moves;
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_EQ(moves.size(), 20);
    EXPECT_TRUE(contains(moves, Move{Sq::b1, Sq::c3, MoveType::Normal}));
    EXPECT_TRUE(contains(moves, Move{Sq::b1, Sq::a3, MoveType::Normal}));
@@ -48,7 +48,7 @@ TEST(MoveGeneratorTest, CandidateMoves1)
    exec.move(Move{Sq::g1, Sq::h3, MoveType::Normal});
 
    moves.clear();
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_EQ(moves.size(), 20);
    EXPECT_TRUE(contains(moves, Move{Sq::b8, Sq::c6, MoveType::Normal}));
    EXPECT_TRUE(contains(moves, Move{Sq::b8, Sq::a6, MoveType::Normal}));
@@ -78,11 +78,11 @@ TEST(MoveGeneratorTest, CandidateMoves2)
    Board b;
    State s;
    setBoard1(b, s);
-   MoveGenerator gen(b,s);
+   std::vector<Move> moves;
+   MoveGenerator gen(b, s, std::back_inserter(moves));
    MoveExecutor exec(b, s);
 
-   std::vector<Move> moves;
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_EQ(moves.size(), 34);
    EXPECT_TRUE(contains(moves, Move{Sq::d1, Sq::c1, MoveType::Normal}));
    EXPECT_TRUE(contains(moves, Move{Sq::d1, Sq::b1, MoveType::Normal}));
@@ -122,7 +122,7 @@ TEST(MoveGeneratorTest, CandidateMoves2)
    exec.move(Move{Sq::a7, Sq::c7, MoveType::Normal, Piece::Queen});
 
    moves.clear();
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_EQ(moves.size(), 31);
    EXPECT_TRUE(contains(moves, Move{Sq::g4, Sq::f5, MoveType::Normal}));
    EXPECT_TRUE(contains(moves, Move{Sq::g4, Sq::e6, MoveType::Normal}));
@@ -163,11 +163,11 @@ TEST(MoveGeneratorTest, CandidateMoves3)
    Board b;
    State s;
    setBoard2(b, s);
-   MoveGenerator gen(b,s);
+   std::vector<Move> moves;
+   MoveGenerator gen(b, s, std::back_inserter(moves));
    MoveExecutor exec(b, s);
 
-   std::vector<Move> moves;
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_EQ(moves.size(), 17);
    EXPECT_TRUE(contains(moves, Move{Sq::d7, Sq::c7, MoveType::Normal}));
    EXPECT_TRUE(contains(moves, Move{Sq::d7, Sq::c8, MoveType::Normal}));
@@ -190,7 +190,7 @@ TEST(MoveGeneratorTest, CandidateMoves3)
    exec.move(Move{Sq::e4, Sq::f3, MoveType::EnPassant, Piece::Pawn});
 
    moves.clear();
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_EQ(moves.size(), 20);
    EXPECT_TRUE(contains(moves, Move{Sq::b3, Sq::c3, MoveType::Normal}));
    EXPECT_TRUE(contains(moves, Move{Sq::b3, Sq::d3, MoveType::Normal}));
@@ -220,11 +220,11 @@ TEST(MoveGeneratorTest, CandidateMoves4)
    Board b;
    State s;
    setBoard3(b, s);
-   MoveGenerator gen(b,s);
+   std::vector<Move> moves;
+   MoveGenerator gen(b, s, std::back_inserter(moves));
    MoveExecutor exec(b, s);
 
-   std::vector<Move> moves;
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_EQ(moves.size(), 22);
    EXPECT_TRUE(contains(moves, Move{Sq::d8, Sq::e8, MoveType::Normal}));
    EXPECT_TRUE(contains(moves, Move{Sq::d8, Sq::e7, MoveType::Normal}));
@@ -252,7 +252,7 @@ TEST(MoveGeneratorTest, CandidateMoves4)
    exec.move(Move{Sq::b2, Sq::a1, MoveType::KnightPromotion, Piece::Rook});
 
    moves.clear();
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_EQ(moves.size(), 14);
    EXPECT_TRUE(contains(moves, Move{Sq::g7, Sq::g8, MoveType::KnightPromotion}));
    EXPECT_TRUE(contains(moves, Move{Sq::g7, Sq::g8, MoveType::BishopPromotion}));
@@ -276,17 +276,17 @@ TEST(MoveGeneratorTest, BlockedCastles1)
    Board b;
    State s;
    setCastleBoard1(b, s);
-   MoveGenerator gen(b,s);
+   std::vector<Move> moves;
+   MoveGenerator gen(b, s, std::back_inserter(moves));
    MoveExecutor exec(b, s);
 
-   std::vector<Move> moves;
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_FALSE(contains(moves, Move{Sq::e8, Sq::g8, MoveType::ShortCastle}));
    EXPECT_FALSE(contains(moves, Move{Sq::e8, Sq::c8, MoveType::LongCastle}));
 
    exec.move(Move{Sq::a8, Sq::a4, MoveType::Normal});
    moves.clear();
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_FALSE(contains(moves, Move{Sq::e1, Sq::g1, MoveType::ShortCastle}));
    EXPECT_FALSE(contains(moves, Move{Sq::e1, Sq::c1, MoveType::LongCastle}));
 }
@@ -297,17 +297,17 @@ TEST(MoveGeneratorTest, BlockedCastles2)
    Board b;
    State s;
    setCastleBoard2(b, s);
-   MoveGenerator gen(b,s);
+   std::vector<Move> moves;
+   MoveGenerator gen(b, s, std::back_inserter(moves));
    MoveExecutor exec(b, s);
 
-   std::vector<Move> moves;
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_FALSE(contains(moves, Move{Sq::e1, Sq::g1, MoveType::ShortCastle}));
    EXPECT_TRUE(contains(moves, Move{Sq::e1, Sq::c1, MoveType::LongCastle}));
 
    exec.move(Move{Sq::f4, Sq::g3, MoveType::Normal});
    moves.clear();
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_FALSE(contains(moves, Move{Sq::e8, Sq::g8, MoveType::ShortCastle}));
    EXPECT_TRUE(contains(moves, Move{Sq::e8, Sq::c8, MoveType::LongCastle}));
 }
@@ -318,15 +318,15 @@ TEST(MoveGeneratorTest, BlockedCastles3)
    Board b;
    State s;
    setCastleBoard3(b, s);
-   MoveGenerator gen(b,s);
+   std::vector<Move> moves;
+   MoveGenerator gen(b, s, std::back_inserter(moves));
    MoveExecutor exec(b, s);
 
    ThreatGenerator tgen(b, s);
    EXPECT_FALSE(tgen.isInCheck(Colour::White));
    EXPECT_TRUE(tgen.isInCheck(Colour::Black));
 
-   std::vector<Move> moves;
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_FALSE(contains(moves, Move{Sq::e1, Sq::g1, MoveType::ShortCastle}));
    EXPECT_FALSE(contains(moves, Move{Sq::e1, Sq::c1, MoveType::LongCastle}));
 
@@ -335,7 +335,7 @@ TEST(MoveGeneratorTest, BlockedCastles3)
    EXPECT_FALSE(tgen.isInCheck(Colour::Black));
 
    moves.clear();
-   gen(std::back_inserter(moves));
+   gen();
    EXPECT_FALSE(contains(moves, Move{Sq::e8, Sq::g8, MoveType::ShortCastle}));
    EXPECT_FALSE(contains(moves, Move{Sq::e8, Sq::c8, MoveType::LongCastle}));
 }
@@ -393,29 +393,28 @@ TEST(MoveGeneratorTest, Outputting)
 {
    Board b;
    State s;
-   MoveGenerator gen(b,s);
 
    std::vector<Move> moves;
-   gen(std::back_inserter(moves));
+   MoveGenerator(b, s, std::back_inserter(moves))();
    EXPECT_EQ(moves.size(), 20);
 
    std::vector<Move> moves2(20);
-   gen(moves2.begin());
+   MoveGenerator(b, s, moves2.begin())();
    EXPECT_EQ(moves2, moves);
 
    std::list<MoveInt> moveints;
-   gen(std::front_inserter(moveints));
+   MoveGenerator(b, s, std::front_inserter(moveints))();
    EXPECT_EQ(moveints.size(), 20);
 
    MoveInt movearr[20];
-   gen(movearr);
+   MoveGenerator(b, s, movearr)();
    EXPECT_EQ(movearr[19], moveints.front());
 
    std::deque<MoveAN> movestrs;
-   gen(std::insert_iterator(movestrs, movestrs.begin()));
+   MoveGenerator(b, s, std::insert_iterator(movestrs, movestrs.begin()))();
    EXPECT_EQ(movestrs.size(), 20);
 
    std::ostringstream movestream;
-   gen(std::ostream_iterator<MoveAN>(movestream));
+   MoveGenerator(b, s, std::ostream_iterator<MoveAN>(movestream))();
    EXPECT_EQ(movestream.str().size(), 4*20);
 }
